+++ Products/CMFPlone/browser/exceptions.py
@@ -3,7 +3,6 @@ from AccessControl import getSecurityManager
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zExceptions.ExceptionFormatter import format_exception
-from zope.component.hooks import getSite
 
 import json
 import sys
@@ -24,7 +23,7 @@ class ExceptionView(BrowserView):
         error_type = exception.__class__.__name__
         exc_type, value, traceback = sys.exc_info()
         error_tb = ''.join(
-            format_exception(exc_type, value, traceback, as_html=True))
+            format_exception(exc_type, value, traceback, as_html=False))
         request.response.setStatus(exc_type)
 
         # Indicate exception as JSON
@@ -34,24 +33,16 @@ class ExceptionView(BrowserView):
                 'error_type': error_type,
             })
 
-        if getSite() is None:
-            # We cannot get the site, so we cannot render our nice template
-            template = self.basic_template
-        else:
-            # Use a simplified template if main_template is not available
-            try:
-                self.context.unrestrictedTraverse('main_template')
-            except:
-                template = self.basic_template
-            else:
-                template = self.index
-
         # Render page with user-facing error notice
         request.set('disable_border', True)
         request.set('disable_plone.leftcolumn', True)
         request.set('disable_plone.rightcolumn', True)
 
-        return template(
-            error_type=error_type,
-            error_tb=error_tb,
-        )
+        try:
+            return self.index(
+                error_type=error_type,
+                error_tb=error_tb)
+        except:
+            return self.basic_template(
+                error_type=error_type,
+                error_tb=error_tb)